% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demViz.R
\docType{package}
\name{demViz}
\alias{demViz}
\title{Data visualization functions for demography, public health and the
social sciences}
\description{
Visualization functions for demographic data and estimates in
the fields of demography, public health and the social sciences.
}
