% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_graph.R
\name{plot_line_graph}
\alias{plot_line_graph}
\title{Create basic ggplot2 line graph}
\usage{
plot_line_graph(
  dt,
  plot_title,
  x_col,
  x_axis_title,
  y_col,
  y_axis_title,
  y_axis_commas = FALSE,
  y_lb = NULL,
  y_ub = NULL,
  ribbon_alpha = 0.2,
  point_col = NULL,
  facet_type = "grid",
  facet_form = "age_name ~ sex",
  facet_scales = "free_y",
  scale_manual = NULL
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
The data.table with the values to be plotted.}

\item{plot_title}{[\code{character(1)}]\cr
Overall title of the plot.}

\item{x_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to x-value of each line.}

\item{x_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to y-value of each line.}

\item{y_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_axis_commas}{[\code{logical(1)}]\cr
Whether the y-axis scale labels should use comma formatting
(scales::comma). Default is FALSE.}

\item{y_lb}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to the lower bound of the
\code{y_col} variable. Default is NULL and is not plotted.}

\item{y_ub}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to the upper bound of the
\code{y_col} variable. Default is NULL and is not plotted.}

\item{ribbon_alpha}{[\code{numeric(1)}]\cr
The alpha value for the \code{ggplot2::geom_ribbon} call. Default is '0.2'.}

\item{point_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that differentiates between values that should
be plotted as a line versus a point. This column should be made up of
logical values, with 'TRUE' indicating a point.}

\item{facet_type}{[\code{character(1)}]\cr
Whether to use \code{gplot2::facet_wrap} ('wrap') or \code{gplot2::facet_grid}
('grid'). Default is 'grid'. Can be NULL if no faceting is required.}

\item{facet_form}{[\code{character(1)}]\cr
String representing the formula that defines the rows and/or columns in the
facet call.}

\item{facet_scales}{[\code{character(1)}]\cr
\code{scales} argument to \code{gplot2::facet_wrap} or \code{gplot2::facet_grid}. Should
scales be 'fixed, 'free', or free in one dimension ('free_x', 'free_y')?
Default is 'free_y'.}

\item{scale_manual}{[\code{list(1)}]\cr
Defines the manual accessory aesthetics to be used when plotting the lines
(and points). See \strong{Details}.}
}
\value{
[\code{ggplot()}] object
}
\description{
Create a basic ggplot2 line graph with the option to also add
points. This could be used for plotting time series or age patterns for
different parameters. One could just use their own ggplot syntax to make
these graphs but this function provides a basic structure that encourages
improved formatting (requires axis titles, uses commas for the y-axis values
etc.)
}
\details{
The \code{scale_manual} argument is a nested list of lists for each manual
aesthetic applied. The first level of the list is the name of the aesthetic,
the possible aethetics are 'colour', 'fill', 'alpha', 'shape', 'linetype',
and 'size'. Each aesthetic must have a nested list that specifies the column
in \code{dt} that the aesthetic corresponds to (\code{col}), the legend title for the
aesthetic (\code{legend_title}), and the scale that maps values of the aesthetic
to the values of the column in \code{dt} (\code{scale}).

The groupings for lines and points will be determined by the \code{scale_manual}
columns since each combination of those should be plotted separately.
}
\examples{
plot_line_graph(
  dt = demCore::burkina_faso_initial_estimates$asfr,
  plot_title = "Burknia Faso ASFR",
  x_col = "year_start",
  x_axis_title = "Start of Year Interval",
  y_col = "value",
  y_axis_title = "ASFR",
  facet_form = "age_start ~ ."
)

}
\seealso{
\code{vignette("plot_line_graph")}
}
