% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create basic ggplot2 scatter}
\usage{
plot_scatter(
  dt,
  plot_title,
  x_col,
  x_axis_title,
  y_col,
  y_axis_title,
  y_axis_commas = FALSE,
  equivalence_line = FALSE,
  facet_type = "grid",
  facet_form = "age_name ~ sex",
  facet_scales = "free_y",
  scale_manual = NULL
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
The data.table with the values to be plotted.}

\item{plot_title}{[\code{character(1)}]\cr
Overall title of the plot.}

\item{x_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to x-value.}

\item{x_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to y-value.}

\item{y_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_axis_commas}{[\code{logical(1)}]\cr
Whether the y-axis scale labels should use comma formatting
(scales::comma). Default is FALSE.}

\item{equivalence_line}{[\code{logical(1)}]\cr
Whether to add a line with slope 1 and intercept 0 to show where x = y.
Default FALSE.}

\item{facet_type}{[\code{character(1)}]\cr
Whether to use \code{gplot2::facet_wrap} ('wrap') or \code{gplot2::facet_grid}
('grid'). Default is 'grid'. Can be NULL if no faceting is required.}

\item{facet_form}{[\code{character(1)}]\cr
String representing the formula that defines the rows and/or columns in the
facet call.}

\item{facet_scales}{[\code{character(1)}]\cr
\code{scales} argument to \code{gplot2::facet_wrap} or \code{gplot2::facet_grid}. Should
scales be 'fixed, 'free', or free in one dimension ('free_x', 'free_y')?
Default is 'free_y'.}

\item{scale_manual}{[\code{list(1)}]\cr
Defines the manual accessory aesthetics to be used when plotting the lines
(and points). See \strong{Details}.}
}
\value{
[\code{ggplot()}] object
}
\description{
Create a basic ggplot2 scatter plot with the option to also add
a line of equivalence. This could be used for comparing two versions of
the same model or a model and a comparator from another research group. One
could just use their own ggplot syntax to make these graphs but this
function provides a basic structure that encourages improved formatting
(requires axis titles, uses commas for the y-axis values etc.)
}
\details{
The \code{scale_manual} argument is a nested list of lists for each manual
aesthetic applied. The first level of the list is the name of the aesthetic,
the possible aethetics are 'colour', 'fill', 'alpha', 'shape', 'linetype',
and 'size'. Each aesthetic must have a nested list that specifies the column
in \code{dt} that the aesthetic corresponds to (\code{col}), the legend title for the
aesthetic (\code{legend_title}), and the scale that maps values of the aesthetic
to the values of the column in \code{dt} (\code{scale}).

The groupings for lines and points will be determined by the \code{scale_manual}
columns since each combination of those should be plotted separately.
}
\examples{
dt1 <- demCore::burkina_faso_initial_estimates$asfr
dt2 <- demCore::thailand_initial_estimates$asfr
dt <- merge(dt1, dt2, by = c("year_start", "year_end", "age_start", "age_end"))
plot_scatter(
  dt = dt,
  plot_title = "ASFR: Burknia Faso vs Thailand",
  x_col = "value.x",
  x_axis_title = "Burkina Faso ASFR",
  y_col = "value.y",
  y_axis_title = "Thailand ASFR",
  facet_form = "age_start ~ ."
)

}
\seealso{
This function follows similar syntax to \code{\link[=plot_line_graph]{plot_line_graph()}}. See
\code{vignette("plot_line_graph")} for details.
}
