% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_plot_args}
\alias{validate_plot_args}
\title{Validate plot arguments}
\usage{
validate_plot_args(
  dt,
  plot_title,
  x_col,
  x_axis_title,
  y_col,
  y_axis_title,
  y_axis_commas = NULL,
  y_lb = NULL,
  y_ub = NULL,
  ribbon_alpha = NULL,
  point_col = NULL,
  equivalence_line = NULL,
  facet_type = NULL,
  facet_form = NULL,
  facet_scales = NULL,
  scale_manual = NULL
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
The data.table with the values to be plotted.}

\item{plot_title}{[\code{character(1)}]\cr
Overall title of the plot.}

\item{x_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to x-value of each line.}

\item{x_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to y-value of each line.}

\item{y_axis_title}{[\code{character(1)}]\cr
X-axis title.}

\item{y_axis_commas}{[\code{logical(1)}]\cr
Whether the y-axis scale labels should use comma formatting
(scales::comma). Default is FALSE.}

\item{y_lb}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to the lower bound of the
\code{y_col} variable. Default is NULL and is not plotted.}

\item{y_ub}{[\code{character(1)}]\cr
Name of the column in \code{dt} that corresponds to the upper bound of the
\code{y_col} variable. Default is NULL and is not plotted.}

\item{ribbon_alpha}{[\code{numeric(1)}]\cr
The alpha value for the \code{ggplot2::geom_ribbon} call. Default is '0.2'.}

\item{point_col}{[\code{character(1)}]\cr
Name of the column in \code{dt} that differentiates between values that should
be plotted as a line versus a point. This column should be made up of
logical values, with 'TRUE' indicating a point.}

\item{equivalence_line}{[\code{logical(1)}]\cr
Whether to add a line with slope 1 and intercept 0 to show where x = y.
Default FALSE.}

\item{facet_type}{[\code{character(1)}]\cr
Whether to use \code{gplot2::facet_wrap} ('wrap') or \code{gplot2::facet_grid}
('grid'). Default is 'grid'. Can be NULL if no faceting is required.}

\item{facet_form}{[\code{character(1)}]\cr
String representing the formula that defines the rows and/or columns in the
facet call.}

\item{facet_scales}{[\code{character(1)}]\cr
\code{scales} argument to \code{gplot2::facet_wrap} or \code{gplot2::facet_grid}. Should
scales be 'fixed, 'free', or free in one dimension ('free_x', 'free_y')?
Default is 'free_y'.}

\item{scale_manual}{[\code{list(1)}]\cr
Defines the manual accessory aesthetics to be used when plotting the lines
(and points). See \strong{Details}.}
}
\value{
[\code{list()}]\cr aesthetics prepped
}
\description{
Helper function for common validations to demViz plotting
functions. Also includes prep of \code{aesthetics} input.
}
