% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem_palette}
\alias{dem_palette}
\title{Demographics palette}
\format{
A named list, where the names are:
\itemize{
\item "gbd_previous_cycle"
\item "gbd_current"
\item "gbd_recent"
\item "vr"
\item "hmd"
\item "wpp"
\item "stage_1"
\item "stage_2"
\item "prior"
\item "posterior"
\item "data"
}

And the values are hex codes for colors to use for these data types.
}
\usage{
dem_palette
}
\description{
Palette which can be used as \code{values} argument to
\code{ggplot2::scale_color_manual} or \code{ggplot2::scale_fill_manual}, and
includes common GBD comparators and output types.
}
\keyword{datasets}
